* Refactored basic.js * Added common module * Added entity module
