+++ lib/common.js
@@ -1,0 +1,11 @@
+'use strict'
+
+module.exports = {
+  messages: {
+    UTIL_DEPRECATED: 'role:util patterns are replaced by role:basic',
+    MARKED_FOR_REMOVAL: 'marked for removal in future'
+  },
+  plugin: {
+    name: 'basic'
+  }
+}
