+++ lib/note.js
@@ -1,3 +1,5 @@
+'use strict'
+
 module.exports = function () {
   var note_single = {}
   var note_values = {}
