+++ lib/uniqueid.js
@@ -1,3 +1,5 @@
+'use strict'
+
 var nid = require('nid')
 
 function quickcode (args, done) {
